Tune deployment resource use
### Greetings from your friendly neighborhood prbot!

This is an automated pull request that:
 - Sets your cpu.request based on the cost efficiency recommendations that can be found in the [Backstage cost efficiency page] and the cpu.limit to 1.5 times the cpu.request
 - If the cost efficiency recommendations are not available it will set your cpu.request to ~60% of the cpu.limit if it was previously higher and leave cpu.limit as is
 - Sets your memory resource request to the same as the memory limit

Having a lower cpu request than limit is recommended as it allows for higher
utilization in the cluster overall. If you feel this request is too low,
consider increasing the cpu.limit for your service instead.

Changes to the cpu request / limit might require adjustments in the application itself.
For example, changes to thread pool configurations.

For more details on scaling, see the [scaling guide].
For more information around at the cost efficiency of your service and related recommendations see the [Backstage cost efficiency page].

If you feel that this PR is wrong and you'd like to opt out of this type of
automation, add this repository to the [blacklist].

If you have any feedback, please fill out the [feedback form]
or reach out to us in [#kubernetes-users] with any problems or questions.

[scaling guide]: https://backstage.spotify.net/docs/gke/developers/scaling/
[Backstage cost efficiency page]: https://backstage.spotify.net/services/spotify-lemur/efficiency
[blacklist]: https://ghe.spotify.net/alf/k8s-autosuggest/blob/master/blacklist/deployment-limits
[feedback form]: https://docs.google.com/forms/d/e/1FAIpQLSdHqE-lYMkaTr98GMOOtG6ewdKj3k2AJ1MwQpQSQ18vhhJRHw/viewform?usp=pp_url&entry.110429453=deployment-limits
[#kubernetes-users]: https://spotify.slack.com/messages/C5CV494H5