Fix secrets being constantly re-uploaded as new versions
SecretManager assumes that certificates have a particular name and
new versions of that certificate will have exactly the same name.
spotify-lemur does not make that assumption, it gives every cert a
unique name even if they are logically just new versions of the same
cert (e.g. `spotify-com-20220517-d8c301`, `spotify-com-20230517-adadad`).

Due to this mistmatch between the understanding of the significance
of the names the plugin was uploading each cert every 10 minutes
as a new version of the same cert (with the same name). In our set up
two certs with the same name are exactly the same and not possibly
different versions in the rotation chain. This caused "new" versions
of every cert to be added every 10 minutes when really it was the
exact same cert (bytes) being uploaded again and again.

This method previously returned version.name which is in the format
`projects/*/secrets/*/versions/*`, but this was unused anywhere else
in the code and as we are not uploading new versions of the cert
we are not getting a new version name each time. The method has been
changed to return the name of the overall cert `projects/*/secrets/*`
- as it was unused anywhere it shouldn't make a difference, it just
made this change easier.